+++ gui/qt/main_window.cpp
@@ -594,21 +594,21 @@ void main_window::set_feedback_wraparound(bool value)
   set_check_box(feedback_wraparound_checkbox, value);
 }
 
-void main_window::set_pid_proportional(uint16_t multiplier, uint16_t exponent)
+void main_window::set_pid_proportional(uint16_t multiplier, uint8_t exponent)
 {
   suppress_events = true;
   pid_proportional_groupbox->set_values(multiplier, exponent);
   suppress_events = false;
 }
 
-void main_window::set_pid_integral(uint16_t multiplier, uint16_t exponent)
+void main_window::set_pid_integral(uint16_t multiplier, uint8_t exponent)
 {
   suppress_events = true;
   pid_integral_groupbox->set_values(multiplier, exponent);
   suppress_events = false;
 }
 
-void main_window::set_pid_derivative(uint16_t multiplier, uint16_t exponent)
+void main_window::set_pid_derivative(uint16_t multiplier, uint8_t exponent)
 {
   suppress_events = true;
   pid_derivative_groupbox->set_values(multiplier, exponent);
